chore: ajout fichiers séparés (cassages,double_sdes) pour meilleure structuration
