+++ test/test_cassages.py
@@ -1,14 +1,18 @@
 """ Module de tests des cassages (brutal et astucieux) """
-# pylint: disable=E0401
+# pylint: disable=E0401, C0413
 import sys
 import os
-from sae import (double_sdes_encrypt_text, cassage_brutal, cassage_astucieux,
-                 sdes_decrypt_text, CONST, time)
 
 ROOT = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..")
 sys.path.append(os.path.join(ROOT, "src"))
 
+from cassages import (double_sdes_encrypt_text, cassage_brutal, cassage_astucieux,
+                 sdes_decrypt_text, CONST, time)
+
 def test_cassage_brutal():
+    """
+    Test le temps d'exécution du cassage brutal sur le texte dans les constantes
+    """
     start_time = time.time()
     texte_chiffre = double_sdes_encrypt_text(CONST.PREMIERE_CLE_TEST,
                                              CONST.DEUXIEME_CLE_TEST, CONST.TEXTE_TEST)
@@ -26,6 +30,9 @@ def test_cassage_brutal():
     print("Temps d'exécution cassage brutal :", execution_time, "secondes\n")
 
 def test_cassage_astucieux():
+    """
+    Test le temps d'exécution du cassage astucieux sur le texte dans les constantes
+    """
     start_time = time.time()
     texte_chiffre = double_sdes_encrypt_text(CONST.PREMIERE_CLE_TEST,
                                              CONST.DEUXIEME_CLE_TEST, CONST.TEXTE_TEST)
