+++ test/test_double_sdes.py
@@ -1,0 +1,22 @@
+""" Module de tests de double sdes (déchiffrement et chiffrement) """
+# pylint: disable=E0401, C0413
+import sys
+import os
+
+ROOT = os.path.join(os.path.dirname(os.path.abspath(__file__)), "..")
+sys.path.append(os.path.join(ROOT, "src"))
+
+from cassages import (double_sdes_encrypt_text, CONST, time)
+
+def test_double_sdes():
+    texte_chiffre_1 = sdes_encrypt_text(CONST.PREMIERE_CLE_TEST, CONST.TEXTE_TEST)
+    print("Texte chiffré 1er passage : \n", texte_chiffre_1)
+    texte_chiffre_2 = sdes_encrypt_text(CONST.DEUXIEME_CLE_TEST, texte_chiffre_1)
+    print("Texte chiffré 2e passage : \n", texte_chiffre_2)
+    texte_dechiffre_1 = sdes_decrypt_text(CONST.DEUXIEME_CLE_TEST, texte_chiffre_2)
+    print("Texte déchiffré 1er passage : \n", texte_dechiffre_1)
+    texte_dechiffre_2 = sdes_decrypt_text(CONST.PREMIERE_CLE_TEST, texte_dechiffre_1)
+    print("Texte déchiffré 2e passage : \n", texte_dechiffre_2,"\n")
+
+if __name__ == "__main__":
+    test_double_sdes()
